String to Morse ellaca-f
First I declare an interpreter to run the program. Then I import sys to be able to pass arguments and also re to use compile function from this library.
Now I declare a dictioary with the translation of every word to morse.

In line 38, I test if there are more than 1 arg, and return error string if that's true.
Line 40, error test of a empty string.
Line 42, error test of not alpha chars. I've used a regex function(re.compile) that create a string whith only letter from a-z and A-Z to use it as a pattern to compare sys.argv[1] and be able to know if there is not alpha char.
Finally I print whit a for loop every letter translated to morse using the dictionary and coverted previously to lowercase. Also I add an end of line to the end.